# output/final_deliverables/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nifty 100 Financial Intelligence API",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/health": {
      "get": {
        "tags": [
          "Health"
        ],
        "summary": "Get Health",
        "description": "Return server status, DB row counts, and uptime.",
        "operationId": "get_health_api_v1_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Get Health Api V1 Health Get"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/companies": {
      "get": {
        "tags": [
          "Companies"
        ],
        "summary": "List Companies",
        "description": "Return a list of all companies, optionally filtered by sector.",
        "operationId": "list_companies_api_v1_companies_get",
        "parameters": [
          {
            "name": "sector",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by broad sector",
              "title": "Sector"
            },
            "description": "Filter by broad sector"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/companies/{ticker}": {
      "get": {
        "tags": [
          "Companies"
        ],
        "summary": "Get Company",
        "description": "Return detailed company profile including latest KPIs and valuation.",
        "operationId": "get_company_api_v1_companies__ticker__get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/companies/{ticker}/tearsheet": {
      "get": {
        "tags": [
          "Companies"
        ],
        "summary": "Get Tearsheet",
        "description": "Download the generated PDF tearsheet for a specific company.",
        "operationId": "get_tearsheet_api_v1_companies__ticker__tearsheet_get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/screener": {
      "get": {
        "tags": [
          "Screener"
        ],
        "summary": "Screen Companies",
        "description": "Return ranked company list based on dynamic filter thresholds.",
        "operationId": "screen_companies_api_v1_screener_get",
        "parameters": [
          {
            "name": "min_roe",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Roe"
            }
          },
          {
            "name": "max_de",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max De"
            }
          },
          {
            "name": "min_fcf",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Fcf"
            }
          },
          {
            "name": "sector",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sector"
            }
          },
          {
            "name": "min_rev_cagr_5yr",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Rev Cagr 5Yr"
            }
          },
          {
            "name": "min_pat_cagr_5yr",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Pat Cagr 5Yr"
            }
          },
          {
            "name": "max_pe",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Pe"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Screen Companies Api V1 Screener Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sectors": {
      "get": {
        "tags": [
          "Sectors"
        ],
        "summary": "List Sectors",
        "description": "Return all sectors with company count and median KPIs.",
        "operationId": "list_sectors_api_v1_sectors_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "type": "array",
                  "title": "Response List Sectors Api V1 Sectors Get"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sectors/{sector}/companies": {
      "get": {
        "tags": [
          "Sectors"
        ],
        "summary": "Get Sector Companies",
        "description": "Return all companies in a specific sector.",
        "operationId": "get_sector_companies_api_v1_sectors__sector__companies_get",
        "parameters": [
          {
            "name": "sector",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sector"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Get Sector Companies Api V1 Sectors  Sector  Companies Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/peers/{group_name}": {
      "get": {
        "tags": [
          "Peers"
        ],
        "summary": "Get Peer Group",
        "description": "Return all companies in a peer group with percentile ranks.",
        "operationId": "get_peer_group_api_v1_peers__group_name__get",
        "parameters": [
          {
            "name": "group_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Group Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Get Peer Group Api V1 Peers  Group Name  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/companies/{ticker}/peers/compare": {
      "get": {
        "tags": [
          "Peers"
        ],
        "summary": "Compare Peers",
        "description": "Return radar data: company vs peer group average.",
        "operationId": "compare_peers_api_v1_companies__ticker__peers_compare_get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Compare Peers Api V1 Companies  Ticker  Peers Compare Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/market-cap/{ticker}": {
      "get": {
        "tags": [
          "Valuation"
        ],
        "summary": "Get Market Cap",
        "description": "Return historical valuation multiples for a company.",
        "operationId": "get_market_cap_api_v1_market_cap__ticker__get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Get Market Cap Api V1 Market Cap  Ticker  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/portfolio/stats": {
      "get": {
        "tags": [
          "Portfolio"
        ],
        "summary": "Get Portfolio Stats",
        "description": "Return P10 through P90 percentile table for core KPIs.",
        "operationId": "get_portfolio_stats_api_v1_portfolio_stats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "type": "array",
                  "title": "Response Get Portfolio Stats Api V1 Portfolio Stats Get"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/companies/{ticker}/documents": {
      "get": {
        "tags": [
          "Documents"
        ],
        "summary": "Get Documents",
        "description": "Return annual report links with validation flag.",
        "operationId": "get_documents_api_v1_companies__ticker__documents_get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Get Documents Api V1 Companies  Ticker  Documents Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CompanyBase": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Company ticker symbol"
          },
          "company_name": {
            "type": "string",
            "title": "Company Name",
            "description": "Full company name"
          },
          "broad_sector": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Broad Sector",
            "description": "Broad sector classification"
          }
        },
        "type": "object",
        "required": [
          "id",
          "company_name"
        ],
        "title": "CompanyBase",
        "description": "Base company schema with ID, name, and sector."
      },
      "CompanyDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Company ticker symbol"
          },
          "company_name": {
            "type": "string",
            "title": "Company Name",
            "description": "Full company name"
          },
          "broad_sector": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Broad Sector",
            "description": "Broad sector classification"
          },
          "roe": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Roe",
            "description": "Latest Return on Equity (%)"
          },
          "roce": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Roce",
            "description": "Latest Return on Capital Employed (%)"
          },
          "debt_to_equity": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Debt To Equity",
            "description": "Latest Debt-to-Equity ratio"
          },
          "pe_ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pe Ratio",
            "description": "Latest Price-to-Earnings ratio"
          },
          "market_cap_crore": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market Cap Crore",
            "description": "Latest Market Capitalization (INR Crore)"
          }
        },
        "type": "object",
        "required": [
          "id",
          "company_name"
        ],
        "title": "CompanyDetail",
        "description": "Detailed company schema including latest KPIs and valuation metrics."
      },
      "CompanyList": {
        "properties": {
          "count": {
            "type": "integer",
            "title": "Count",
            "description": "Total number of companies returned"
          },
          "companies": {
            "items": {
              "$ref": "#/components/schemas/CompanyBase"
            },
            "type": "array",
            "title": "Companies",
            "description": "List of company objects"
          }
        },
        "type": "object",
        "required": [
          "count",
          "companies"
        ],
        "title": "CompanyList",
        "description": "Paginated list of companies."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}